Plus Task DV-6
new Bug in StartDate and EndDiate

+++ DevEduEducationSystem.API.Tests/Support/Models/AllGroupsModels/GroupRequestModel.cs
@@ -28,6 +28,6 @@ namespace DevEduEducationSystem.API.Tests.Support.Models
         public string Timetable { get; set; }
 
         [JsonPropertyName("paymentPerMonth")]
-        public int PaymentPerMonth { get; set; }
+        public decimal PaymentPerMonth { get; set; }
     }
 }
